requirements.txt
chess
python-dotenv
stockfish